\name{firstpaket-package}
\alias{firstpaket-package}
\alias{firstpaket}
\docType{package}
\title{
\packageTitle{firstpaket}
}
\description{
\packageDescription{firstpaket}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{firstpaket}
\packageIndices{firstpaket}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{firstpaket}

Maintainer: \packageMaintainer{firstpaket}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
x<-matrix(c(1:9),nrow=3)
leu(x)
}